歌詞/cl.lyrics-master/requirements.txt
scrapy
python-dotenv
bs4